+++ domain/utils/activity_utils.py
@@ -1,7 +1,7 @@
 from domain.enums.activity_type import ActivityType
 import random
 
-FREE_ACTIVITIES = [ActivityType.PRACTICE, ActivityType.LABOUR, ActivityType.IDLE]
+FREE_ACTIVITIES = [ActivityType.PRACTICE, ActivityType.LABOUR, ActivityType.IDLE, ActivityType.MINING]
 
 
 def choose_activity(extra_activities: list[ActivityType]) -> ActivityType: